stats-service refactor + added schema to /common
